add class service

+++ src/main/java/fall24/swp391/KoiOrderingSystem/repo/IBookingKoiDetailRepository.java
@@ -3,5 +3,6 @@ package fall24.swp391.KoiOrderingSystem.repo;
 import fall24.swp391.KoiOrderingSystem.pojo.BookingKoiDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IBookingKoiDetailRepository extends JpaRepository<BookingKoiDetail, Integer>{
+
+public interface IBookingKoiDetailRepository extends JpaRepository<BookingKoiDetail, Long> {
 }
